requirements.txt
dash==2.9.2
pyyaml
gunicorn
pandas==2.0.0
plotly==5.14.1